Servico por tecnico implementado, realizando melhoras

+++ src/main/java/br/com/w4solution/controle_instalacao/dto/registro/ServicosEquipe.java
@@ -1,4 +1,4 @@
 package br.com.w4solution.controle_instalacao.dto.registro;
 
-public record ServicosEquipe(String label, Integer value) {
+public record ServicosEquipe(String label, Long value) {
 }
